Fixed broken home shortcuts

+++ omniNotes/src/main/java/it/feio/android/omninotes/DetailFragment.java
@@ -401,7 +401,7 @@ public class DetailFragment extends BaseFragment implements OnReminderPickedList
 		if (Constants.ACTION_SHORTCUT.equals(i.getAction())
 				|| Constants.ACTION_NOTIFICATION_CLICK.equals(i.getAction())) {
 			afterSavedReturnsToList = false;
-			noteOriginal = DbHelper.getInstance().getNote(i.getIntExtra(Constants.INTENT_KEY, 0));
+			noteOriginal = DbHelper.getInstance().getNote(i.getLongExtra(Constants.INTENT_KEY, 0));
 			// Checks if the note pointed from the shortcut has been deleted
 			try {
 				note = new Note(noteOriginal);
